* master: (55 commits)
  extract deprecated dynamic methods
  Add some docs and changelog entry
  Allow overriding exception handling in threaded consumer
  Allow configuring a different queue consumer
  actually don't need to expand the aggregates at all
  #to_sym is unnecessary
  de-globalise method
  extract code from AR::Base
  clean up implementation of dynamic methods. use method compilation etc.
  Fix ActiveModel README example
  mention database mapping in getting started guide
  Remove vestiges of the http_only! config from configuring guide
  Remove content-length as well
  Make ActionController#head pass rack-link
  RouteSet: optimize routes generation when globbing is used
  Allows assert_redirected_to to accept a regular expression
  use extract_options!
  No need to force conversion to Symbol since case ensures it's already one.
  No need to work around 1.8 warnings anymore.
  Update command line guide
  ...

+++ activesupport/lib/active_support/cache.rb
@@ -280,7 +280,7 @@ module ActiveSupport
             end
           end
           if entry && entry.expired?
-            race_ttl = options[:race_condition_ttl].to_f
+            race_ttl = options[:race_condition_ttl].to_i
             if race_ttl and Time.now.to_f - entry.expires_at <= race_ttl
               entry.expires_at = Time.now + race_ttl
               write_entry(key, entry, :expires_in => race_ttl * 2)
